refactor: WIP first pass at updateMediaMetadata function, commented out the media.before.update hook
