+++ imports/plugins/core/files/server/methods.js
@@ -1,7 +1,68 @@
 import { Meteor } from "meteor/meteor";
 import { check } from "meteor/check";
+import includes from "lodash/includes";
+import { Hooks } from "/server/api";
 import { Reaction } from "/server/api";
-import { MediaRecords } from "/lib/collections";
+import { MediaRecords, Revisions } from "/lib/collections";
+import { RevisionApi } from "/imports/plugins/core/revisions/lib/api";
+
+/**
+ * @method updateMediaMetadata
+ * @memberof media
+ * @summary insert a new media record and add it to revision control.
+ * @param {Object} fileRecord - document from file collection upload.
+ * @return {String} - _id of the new inserted media record.
+ */
+export async function updateMediaMetadata(fileRecordId, metadata) {
+  console.log("Update Media Metadata", fileRecordId, metadata)
+  check(fileRecordId, String);
+  check(metadata, Object);
+  if (RevisionApi.isRevisionControlEnabled()) {
+    if (metadata.productId) {
+      const existingRevision = Revisions.findOne({
+        "documentId": fileRecordId,
+        "workflow.status": {
+          $nin: [
+            "revision/published"
+          ]
+        }
+      });
+      if (existingRevision) {
+        const updatedMetadata = Object.assign({}, existingRevision.documentData, metadata);
+        // Special case where if we have both added and reordered images before publishing we don't want to overwrite
+        // the workflow status since it would be "unpublished"
+        if (existingRevision.documentData.workflow === "published" || existingRevision.changeType === "insert") {
+          updatedMetadata.workflow = "published";
+        }
+        Revisions.update({ _id: existingRevision._id }, {
+          $set: {
+            documentData: updatedMetadata
+          }
+        });
+        Hooks.Events.run("afterRevisionsUpdate", Meteor.userId(), {
+          ...existingRevision,
+          documentData: updatedMetadata
+        });
+      } else {
+        Revisions.insert({
+          documentId: fileRecordId,
+          documentData: metadata,
+          documentType: "image",
+          parentDocument: metadata.productId,
+          changeType: "update",
+          workflow: {
+            status: "revision/update"
+          }
+        });
+      }
+
+      return false; // prevent actual update of image. This also stops other hooks from running :/
+    }
+  }
+  // for non-product images, just ignore and keep on moving
+  return true;
+}
+
 
 Meteor.methods({
   /**
@@ -43,6 +104,8 @@ Meteor.methods({
           "metadata.priority": index
         }
       });
+      const { metadata } = MediaRecords.findOne({ _id });
+      updateMediaMetadata(_id, metadata);
     });
 
     return true;
